requirements.txt
numpy
Pillow
torch>=1.7
torchvision>=0.8.0
tqdm
huggingface-hub
leb128==1.0.4